refactor : AnimatedRouletteWheel 컴포넌트로 따로 분리
